+++ src/entity/interfaces/Enemies.java
@@ -2,7 +2,7 @@ package entity.interfaces;
 
 public abstract class Enemies extends Unit{
 	
-	public Enemies(String name, int health, int attackPower, double speed, double range, boolean isAlley, int accuracy,int evasion , double cooldown) {
+	public Enemies(String name, int health, int attackPower, int speed, double range, boolean isAlley, int accuracy,int evasion , double cooldown) {
 		super(name, health, attackPower, speed, range, isAlley, accuracy, evasion , cooldown);
 		this.setPos(640);
 	}
